Fixed issue #18961: {SURVEYURL} not complete URL
